requirements.txt
# Required for the included local-text demo: Python standard library only.
#
# Optional tests/API/document parsing/OCR packages:
pytest
fastapi
uvicorn
python-multipart
psycopg[binary]
pymupdf
pillow
rapidocr-onnxruntime
# pytesseract
# docling
requests
python-dotenv
